[PAGOPA-931] fix: renamed package
